refactor: rename private nplikes & `NumpyModuleArray`
